Type 및 전반적인 오류 수정

+++ src/main/java/com/swacademy/mapcommunity/domain/repository/CommentRepository.java
@@ -3,5 +3,6 @@ package com.swacademy.mapcommunity.domain.repository;
 import com.swacademy.mapcommunity.domain.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends JpaRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Integer> {
+
 }
